+++ mapie/metrics.py
@@ -1,9 +1,9 @@
-from sklearn.utils.validation import column_or_1d
-
+from sklearn.utils.validation import column_or_1d, check_array
+import numpy as np
 from ._typing import ArrayLike
 
 
-def coverage_score(
+def regression_coverage_score(
     y_true: ArrayLike,
     y_pred_low: ArrayLike,
     y_pred_up: ArrayLike,
@@ -30,16 +30,63 @@ def coverage_score(
 
     Examples
     --------
-    >>> from mapie.metrics import coverage_score
+    >>> from mapie.metrics import regression_coverage_score
     >>> import numpy as np
     >>> y_true = np.array([5, 7.5, 9.5, 10.5, 12.5])
     >>> y_pred_low = np.array([4, 6, 9, 8.5, 10.5])
     >>> y_pred_up = np.array([6, 9, 10, 12.5, 12])
-    >>> print(coverage_score(y_true, y_pred_low, y_pred_up))
+    >>> print(regression_coverage_score(y_true, y_pred_low, y_pred_up))
     0.8
     """
     y_true = column_or_1d(y_true)
     y_pred_low = column_or_1d(y_pred_low)
     y_pred_up = column_or_1d(y_pred_up)
     coverage = ((y_pred_low <= y_true) & (y_pred_up >= y_true)).mean()
+    return float(coverage)
+
+
+def classification_coverage_score(
+    y_true: ArrayLike,
+    y_pred_set: ArrayLike
+) -> float:
+    """
+    Effective coverage score obtained by the prediction sets.
+
+    The effective coverage is obtained by estimating the fraction
+    of true labels that lie within the prediction sets.
+
+    Parameters
+    ----------
+    y_true : ArrayLike of shape (n_samples,)
+        True labels.
+    y_pred_set : ArrayLike of shape (n_samples, n_class)
+        Prediction sets given by booleans of labels.
+
+    Returns
+    -------
+    float
+        Effective coverage obtained by the prediction sets.
+
+    Examples
+    --------
+    >>> from mapie.metrics import classification_coverage_score
+    >>> import numpy as np
+    >>> y_true = np.array([3, 3, 1, 2, 2])
+    >>> y_pred_set = np.array([
+    ...     [False, False,  True,  True],
+    ...     [False,  True, False,  True],
+    ...     [False,  True,  True, False],
+    ...     [False, False,  True,  True],
+    ...     [False,  True, False,  True]
+    ... ])
+    >>> print(classification_coverage_score(y_true, y_pred_set))
+    0.8
+    """
+    y_true = column_or_1d(y_true)
+    y_pred_set = check_array(
+        y_pred_set, force_all_finite=True, dtype=["bool"]
+    )
+    coverage = np.take_along_axis(
+        y_pred_set, y_true.reshape(-1, 1), axis=1
+    ).mean()
     return float(coverage)