+++ HISTORY.rst
@@ -2,11 +2,19 @@
 History
 =======
 
-0.2.4 (XXXX-XX-XX)
+0.3.0 (XXXX-XX-XX)
 ------------------
 
 * Renaming estimators.py module to regression.py
-* Modification of the documentation architecture 
+* New classification.py module with MapieClassifier class, that estimates prediction sets from softmax score
+* New set of unit tests for classification.py module
+* Modification of the documentation architecture
+* Split example gallery into separate regression and classification galleries
+* Add first classification examples
+* Add method classification_coverage_score in the module metrics.py
+* Fixed code error for plotting of interval widths in tutorial of documentation
+* Added missing import statements in tutorial of documentation
+* Refactorize tests of `n_jobs` and `verbose` in `utils.py`
 
 0.2.3 (2021-07-09)
 ------------------